CB-14728 Enable stopping clusters with ephemeral based hostgroups based on service role whitelist.
This feature enables stopping clusters with ephemeral volume storage based instance groups. Currently stopping clusters with ephemeral isntance groups is not permitted because in the event of stopping the nodes the ephemeral volumes are formatted, which can cause problems with service components, like HDFS Datanode or Namenode, that are not designed to survive such data loss.
However thare are certain service components, like the YARN Nodemanager, that do not store persistent data on the volumes. These service components can work corretly if the ephemeral volume under them is formatted and the data is lost, as they do not require persistency.

The aim of the feature is to extend the current stop restriction logic to check ephemeral volume based instance groups if the service components on them are in the group of service components that do not require persistency. These service components (or sometimes called service roles) and wildcarded components/roles (where the service is itself is not relevant, only the component/role, like all the GATEWAY component/roles of different services) are enumerated as a configuration.
The service components from groups (so it can be extended in the future with more complex logic) in the configuration. Each "permitted group" contains several service components and wildcarded components and is basically treated as a whitelist. If an ephemeral instance group only contains service components that mach a "permitted group" than the instance group is considered stoppable. As an additional restriction every element in the group can be set to required, which means that these service components has to be present on the instance group for it to be considered stoppable.

The StopRestrictionService iterates over each instance group and check each of them if they are stoppable. If an instance group is not stoppable then it returns with the restriction reason, otherwise permits the stop operation. If the currently inspected instance group is ephemeral based (all volumes on the instance group are ephemeral type or there is at least one ephemeral volume and all other non-ephemeral volumes are used for database) than first the service checks if the cluster was started after this feature is introduced (via CB version) or the salt scripts are upgraded to the required CB version. If the versions are older then the isntance group is not stoppable. Then, if the version is correct then the service iterates over the "permitted groups" and checks if the isntance group's service components mach a group. If a match is found then the instance group is stoppable and the service moves to the next isntance group otherwise the instance group is not stoppable. The logic for non-ephemeral instance group is not changed.

Currently there is onlt one general whitelist group, which contains the YARN NODEMANAGER service role that is not required and a GATEWAY role that is also not required. For example the default DataEngineering templates' "compute" group matches these criterias.